+++ CHANGELOG.md
@@ -5,6 +5,9 @@ All notable changes to this project will be documented in this file.
 ### Added
 - Added the gitconfig file so the phorge version informaiton can be displayed on the http(s)://<phorgeURL>/config/ page 
 
+### Changed
+- Changed the environment variables for S3 Storage configuration
+
 ## [0.3.1] - 2025-12-11
 ### Changed
 - Updated MariaDB to version 10.11